requirements.txt
django
djangorestframework
mysqlclient
markdown
django-filter
djangorestframework-simplejwt
